Add cluster health check to SMB service start routine
smbd will panic in some circustances if it is unable to establish
a global ctdb messaging context. This may happen while trying to
intiialize a loadparm context with a clustered registry backend
while ctdb is stopped. Avoid this and other possibly complications
by simply refusing to start the SMB service if cluster is unhealthy.